% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BREMSC.R
\name{jointDIMMSC}
\alias{jointDIMMSC}
\title{jointDIMMSC function}
\usage{
jointDIMMSC(
  dataProtein,
  dataRNA,
  K,
  useGene = 100,
  maxiter = 100,
  tol = 1e-04,
  lik.tol = 0.01
)
}
\arguments{
\item{dataProtein}{a D*C matrix with D proteins and C cells}

\item{dataRNA}{a G*C matrix with G genes and C cells}

\item{K}{number of desired clusters}

\item{useGene}{number of genes to keep}

\item{maxiter}{maximum number of iterations, with default value 100}

\item{tol}{a convergence tolerance for the difference of vector pie between iterations, with default value 1e-4}

\item{lik.tol}{a convergence tolerance for the difference of log-likelihoods between iterations, with default value 1e-2}
}
\value{
jointDIMMSC returns a list object containing:

\itemize{
  \item clusterID: estimated label for each cell
  \item posteriorProb: a C*K matrix with probability that each cell belongs to each cluster
  \item logLik: estimated log likelihood
  \item alphaMtxProtein: a K*D matrix of alpha estimates for protein source
  \item alphaMtxRNA: a K*G matrix of alpha estimates for RNA source
}
}
\description{
jointDIMMSC function
}
\examples{
# Load the example data data_DIMMSC
data("dataADT")
data("dataRNA")
# Test run of jointDIMMSC: use small number of MCMC to save time
testRun = jointDIMMSC(dataADT, dataRNA, K=4)
# End
}
\references{
Xinjun Wang, Zhe Sun, Yanfu Zhang, Heng Huang, Kong Chen, Ying Ding, Wei Chen. BREM-SC: A Bayesian Random Effects Mixture Model for Joint Clustering Single Cell Multi-omics Data. Submitted 2019.
}
\author{
Xinjun Wang <xiw119@pitt.edu>, Zhe Sun <zhs31@pitt.edu>, Wei Chen <wei.chen@chp.edu>.
}
