% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BREMSC.R
\name{BREMSC}
\alias{BREMSC}
\alias{jointDIMMSC}
\title{A novel joint clustering method for scRNA-seq and CITE-seq data}
\usage{
BREMSC(
  dataProtein,
  dataRNA,
  K,
  nChains = 3,
  nMCMC = 1000,
  sd_alpha_range = c(0.5, 1.5),
  sd_b_range = c(0.2, 1),
  sigmaB = 0.8
)

jointDIMMSC(
  dataProtein,
  dataRNA,
  K,
  maxiter = 100,
  tol = 1e-04,
  lik.tol = 0.01
)
}
\arguments{
\item{dataProtein}{a D*C matrix with D proteins and C cells}

\item{dataRNA}{a G*C matrix with G genes and C cells}

\item{K}{number of desired clusters}

\item{nChains}{number of MCMC chains to run in parallel, each using a CPU}

\item{nMCMC}{number of MCMCs}

\item{sd_alpha_range}{range of random walk depth for alpha}

\item{sd_b_range}{range of random walk depth for b}

\item{sigmaB}{initial estimate of SD of random effects}

\item{maxiter}{maximum number of iterations, with default value 100}

\item{tol}{a convergence tolerance for the difference of vector pie between iterations, with default value 1e-4}

\item{lik.tol}{a convergence tolerance for the difference of log-likelihoods between iterations, with default value 1e-2}
}
\value{
BREMSC returns a list object containing:

\itemize{
  \item clusterID: estimated label for each cell
  \item posteriorProb: a C*K matrix with probability that each cell belongs to each cluster
  \item sdRF_final: estimated SD of random effects
  \item logLik_final: estimated log likelihood
  \item vecB: estimated random effects for each cell
  \item alphaMtxProtein: a K*D matrix of alpha estimates for protein source
  \item alphaMtxRNA: a K*G matrix of alpha estimates for RNA source
  \item vecLogLik: estimated log likelihood for each MCMC
  \item vecSDRF: estimated SD of random effects for each MCMC
}

jointDIMMSC returns a list object containing:

\itemize{
  \item clusterID: estimated label for each cell
  \item posteriorProb: a C*K matrix with probability that each cell belongs to each cluster
  \item logLik: estimated log likelihood
  \item alphaMtxProtein: a K*D matrix of alpha estimates for protein source
  \item alphaMtxRNA: a K*G matrix of alpha estimates for RNA source
}
}
\description{
BREMSC is developed to joint cluster scRNA-seq and CITE-seq data with the introduction of random effects to incorporate the two data sources

jointDIMMSC is developed as a direct extension of DIMMSC without the introduction of random effects to incorporate the two data sources
}
\examples{
# Load the example data
data("dataADT")
data("dataRNA")
# Test run of jointDIMM-SC
testRun_joint = jointDIMMSC(dataADT, dataRNA, K=4)
# Test run of BREM-SC: use small number of MCMC to save time
testRun_BREMSC = BREMSC(dataADT, dataRNA, K=4, nMCMC = 200)
# Check convergence of log likelihoos
plot(testRun_BREMSC$vecLogLik, type = "l", xlab = "MCMC Iterations", ylab = "Log likelihood")
# End
}
\references{
Xinjun Wang, Zhe Sun, Yanfu Zhang, Zhongli Xu, Hongyi Xin, Heng Huang, Richard H Duerr, Kong Chen, Ying Ding, Wei Chen. BREM-SC: A Bayesian Random Effects Mixture Model for Joint Clustering Single Cell Multi-omics Data. Submitted 2019.
}
\author{
Xinjun Wang <xiw119@pitt.edu>, Zhe Sun <zhs31@pitt.edu>, Wei Chen <wei.chen@chp.edu>.
}
